+++ dots/.config/quickshell/ii/modules/common/functions/Session.qml
@@ -17,11 +17,7 @@ Singleton {
     }
 
     function lock() {
-        if (WM.compositor === "niri") {
-            Quickshell.execDetached(["qs", "-c", "end4-pC", "ipc", "call", "lock", "activate"]);
-        } else {
-            Quickshell.execDetached(["loginctl", "lock-session"]);
-        }
+        Quickshell.execDetached(["loginctl", "lock-session"]);
     }
 
     function suspend() {
@@ -30,11 +26,7 @@ Singleton {
 
     function logout() {
         closeAllWindows();
-        if (WM.compositor === "niri") {
-            Quickshell.execDetached(["niri", "msg", "action", "quit"]);
-        } else {
-            Quickshell.execDetached(["pkill", "-i", "Hyprland"]);
-        }
+        Quickshell.execDetached(["pkill", "-i", "Hyprland"]);
     }
 
     function launchTaskManager() {