+++ dots/.config/quickshell/ii/modules/ii/bar/HyprlandXkbIndicator.qml
@@ -26,7 +26,7 @@ Loader {
             StyledText {
                 id: layoutCodeText
                 horizontalAlignment: Text.AlignHCenter
-                text: WM.compositor === "niri" ? NiriXkb.currentLayoutCode : HyprlandXkb.currentLayoutCode
+                text: HyprlandXkb.currentLayoutCode
                 font.pixelSize: Appearance.font.pixelSize.small
                 color: root.color
                 animateChange: true