+++ dots/.config/quickshell/ii/modules/ii/settings/SettingsContent.qml
@@ -142,10 +142,7 @@ Item {
             { name: Translation.tr("Interface"), icon: "bottom_app_bar", component: Qt.resolvedUrl("pages/InterfaceConfig.qml"), sections: [Translation.tr("Overview"), Translation.tr("Dock"), Translation.tr("Icon pack"), Translation.tr("Lock screen"), Translation.tr("Overlay"), Translation.tr("Region selector"), Translation.tr("On-screen display"), Translation.tr("Wallpaper selector"), Translation.tr("Fonts"), Translation.tr("Terminal"), Translation.tr("Screenshot popup"), Translation.tr("Color generation"), Translation.tr("Keep awake"), Translation.tr("Screensaver")] },
             { name: Translation.tr("Services"), icon: "settings", component: Qt.resolvedUrl("pages/ServicesConfig.qml"), sections: [Translation.tr("AI"), Translation.tr("Networking"), Translation.tr("Music Recognition"), Translation.tr("Screen recorder"), Translation.tr("Save paths"), Translation.tr("Search"), Translation.tr("System updates"), Translation.tr("Weather")] },
         ]
-        if (WM.compositor === "hyprland")
-            list.push({ name: Translation.tr("Hyprland"), icon: "select_window_2", component: Qt.resolvedUrl("pages/HyprlandConfig.qml"), sections: [Translation.tr("Displays"), Translation.tr("Layout"), Translation.tr("Input"), Translation.tr("Visual & Aesthetics"), Translation.tr("Blur"), Translation.tr("Autostart Apps"), Translation.tr("Animations")] })
-        if (WM.compositor === "niri")
-            list.push({ name: Translation.tr("Niri"), icon: "select_window_2", component: Qt.resolvedUrl("pages/NiriConfig.qml"), sections: [Translation.tr("Displays"), Translation.tr("Layout"), Translation.tr("Input"), Translation.tr("Visual & Aesthetics"), Translation.tr("Cursor"), Translation.tr("Animations")] })
+        list.push({ name: Translation.tr("Hyprland"), icon: "select_window_2", component: Qt.resolvedUrl("pages/HyprlandConfig.qml"), sections: [Translation.tr("Displays"), Translation.tr("Layout"), Translation.tr("Input"), Translation.tr("Visual & Aesthetics"), Translation.tr("Blur"), Translation.tr("Autostart Apps"), Translation.tr("Animations")] })
         list.push({ name: Translation.tr("About"), icon: "info", component: Qt.resolvedUrl("pages/About.qml"), sections: [] })
         return list
     }