+++ dots/.config/quickshell/ii/shell.qml
@@ -37,10 +37,6 @@ ShellRoot {
         function onReadyChanged() {
             if (!Config.ready) return
 
-            if (WM.compositor === "niri") {
-                Config.options.background.lockWall = ""
-                Config.options.overview.enable = false
-            }
 
             if (Config.options.hyprland.autostartApps.enable &&
                 Config.options.hyprland.autostartApps.apps.length > 0) {